+++ teamup-client/src/app/components/company/company.component.spec.ts
