+++ teamup-client/src/app/components/industry/industry.component.spec.ts
