+++ teamup-client/src/app/components/industry/industry.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {IndustryService} from "../../services/industry/industry.service";
-import {Industry} from "../../models/industry";
+import { IndustryService } from "../../core/services/industry/industry.service";
 
 @Component({
   selector: 'app-industry',
@@ -9,15 +8,14 @@ import {Industry} from "../../models/industry";
 })
 export class IndustryComponent implements OnInit {
 
-  industries: Industry[];
+  industries: any;
 
   constructor(private industryService: IndustryService) { }
 
   ngOnInit() {
-    this.industryService.getAll().subscribe(
-        data=>{this.industries = data},
-        error1 => {console.log(error1)}
-    );
+    this.industryService.getAll().subscribe((data) => {
+      this.industries = data;
+    });
   }
 
 }
