+++ teamup-client/src/app/components/tech-area/tech-area.component.html
