+++ teamup-client/src/app/components/tech-area/tech-area.component.ts
@@ -1,0 +1,21 @@
+import { Component, OnInit } from '@angular/core';
+import { TechAreaService } from "../../core/services/techArea/tech-area.service";
+
+@Component({
+    selector: 'app-tech-area',
+    templateUrl: './tech-area.component.html',
+    styleUrls: ['./tech-area.component.scss']
+})
+export class TechAreaComponent implements OnInit {
+
+    techAreas: any;
+
+    constructor(private techAreaService: TechAreaService) { }
+
+    ngOnInit() {
+        this.techAreaService.getAll().subscribe((data) => {
+            this.techAreas = data;
+        });
+    }
+
+}
