+++ teamup-client/src/app/core/services/company/company.service.ts
@@ -1,22 +1,18 @@
 import { Injectable } from '@angular/core';
-import {environment} from "../../../../environments/environment";
-import {HttpClient} from "@angular/common/http";
-import {Company} from "../../models/company";
-import {Observable} from "rxjs";
+import { HttpService } from '../../http/http.service';
 
 @Injectable({
-  providedIn: 'root'
+    providedIn: 'root'
 })
 export class CompanyService {
-  baseUrl = environment.serviceHost;
 
-  constructor(private httpClient: HttpClient) { }
+    url: string;
 
-  getAll(): Observable<Company[]>{
-    return this.httpClient.get<Company[]>(this.baseUrl+"company")
-  }
+    constructor(private httpService: HttpService) {
+        this.url = 'company/';
+    }
 
-  save(company: Company){
-    this.httpClient.post<Company>(this.baseUrl+"company/save",company);
-  }
+    getAll() {
+        return this.httpService.get(this.url);
+    }
 }
