+++ teamup-client/src/app/core/services/location/location.service.ts
@@ -1,22 +1,18 @@
 import { Injectable } from '@angular/core';
-import {environment} from "../../../../environments/environment";
-import {HttpClient} from "@angular/common/http";
-import {Location} from "../../models/location";
-import {Observable} from "rxjs";
+import { HttpService } from '../../http/http.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class LocationService {
-  baseUrl = environment.serviceHost;
+  url: string;
 
-  constructor(private httpClient: HttpClient) { }
+    constructor(private httpService: HttpService) {
+        this.url = 'location/';
+    }
 
-  getAll(): Observable<Location[]>{
-    return this.httpClient.get<Location[]>(this.baseUrl+"location")
-  }
+    getAll() {
+        return this.httpService.get(this.url);
+    }
 
-  save(location: Location){
-    this.httpClient.post<Location>(this.baseUrl+"location/save",location);
-  }
 }
