+++ teamup-client/src/app/core/services/techArea/tech-area.service.ts
@@ -1,22 +1,18 @@
 import { Injectable } from '@angular/core';
-import {environment} from "../../../../environments/environment";
-import {HttpClient} from "@angular/common/http";
-import {TechnologyArea} from "../../models/technologyArea";
-import {Observable} from "rxjs";
+import { HttpService } from '../../http/http.service';
 
 @Injectable({
-  providedIn: 'root'
+    providedIn: 'root'
 })
 export class TechAreaService {
-  baseUrl = environment.serviceHost;
+    url: string;
 
-  constructor(private httpClient: HttpClient) { }
+    constructor(private httpService: HttpService) {
+        this.url = 'technology-area/';
+    }
 
-  getAll(): Observable<TechnologyArea[]>{
-    return this.httpClient.get<TechnologyArea[]>(this.baseUrl+"technology-area")
-  }
+    getAll() {
+        return this.httpService.get(this.url);
+    }
 
-  save(techArea: TechnologyArea){
-    this.httpClient.post<TechnologyArea>(this.baseUrl+"technology-area/save",techArea);
-  }
 }
