--- teamup-client/src/app/modules/home.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { HomeComponent } from './home/home.component';
-
-
-
-@NgModule({
-  declarations: [HomeComponent],
-  imports: [
-    CommonModule
-  ]
-})
-export class HomeModule { }
