--- teamup-client/src/app/modules/login.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { SharedModule } from '../shared/shared.module';
-import { LoginComponent } from './login/login.component';
-
-
-@NgModule({
-    imports: [
-        CommonModule,
-        SharedModule,
-    ],
-    declarations: [
-        LoginComponent
-    ]
-})
-export class LoginModule { }
