+++ teamup-client/src/app/shared/models/company.ts
