+++ teamup-client/src/app/shared/models/industry.ts
