+++ teamup-client/src/app/shared/models/location.ts
