+++ teamup-client/src/app/shared/models/project.ts
