+++ teamup-client/src/app/shared/models/technology.ts
