+++ teamup-client/src/app/shared/models/userSeniority.ts
